pyproject.toml
[project]
name = "eu-ai-gdpr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.4.0",
    "docling>=2.68.0",
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.3",
    "langchain-google-genai>=4.2.0",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.0",
    "tiktoken>=0.12.0",
    "torch>=2.9.1",
    "transformers>=4.57.5",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-snapshot>=0.9.0",
    "ruff>=0.14.11",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--cov=src", "--cov-report=term-missing"]


